feat: Update documentation for neurochain
This commit contains an update to the documentation generation scripts to add documentation for neurochain nodes. The script will now collect all documentation for all categories recursively, not only in the main folder as before. It also supports more than two level of child nodes and also documentation in parent modules.

Implements ML-325